requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
anthropic>=0.3.0

# Authentication and security
pyjwt>=2.8.0
python-multipart>=0.0.6

# Caching (optional)
redis>=5.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For TestClient

# Development dependencies
black>=23.9.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.6.0